+++ handlers/CommandHandler.lua
@@ -48,39 +48,6 @@ end
 		- Zone: An entire zone search, doesn't matter which creature in the zone dropped the item. Provides the player with every item that dropped in the provided zone. Partial/full names supported.
 ]]
 
-tbl.Help = function()
-	print(string.format(tbl.L["ADDON_NAME"] .. "%s, %s, %s, %s, %s, %s, %s, %s, %s, %s", tbl.L["COMMAND_DISCORD"], tbl.L["COMMAND_FORMAT"], tbl.L["COMMAND_HISTORY"], tbl.L["CMD_IMPORT"], tbl.L["COMMAND_LOCALE"], tbl.L["COMMAND_LOOT"], tbl.L["COMMAND_MAN"], tbl.L["COMMAND_REMOVE"], tbl.L["COMMAND_SEARCH"], tbl.L["COMMAND_VIEW"]));
+tbl.CheckProgress = function()
+	print(tbl.L["ADDON_NAME"] .. tbl.GetCount(tbl.Items) .. " " .. tbl.L["OF"] .. " " .. tbl.L["APPROXIMATELY"] .. " " .. 154000 .. " (" .. tbl.Round(tbl.GetCount(tbl.Items)/154000) .. "%)")
 end
--- Synopsis: Prints out the available commands to the chat frame.
-
-tbl.Manual = function(args)
-	if args == tbl.L["COMMAND_FORMAT"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_FORMAT"]);
-		print(tbl.L["INFO_MSG_COMMAND_FORMAT"]);
-	elseif args == tbl.L["COMMAND_HELP"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_HELP"]);
-		print(tbl.L["INFO_MSG_COMMAND_HELP"]);
-	elseif args == tbl.L["COMMAND_HISTORY"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_HISTORY"]);
-		print(tbl.L["INFO_MSG_COMMAND_HISTORY"]);
-	elseif args == tbl.L["COMMAND_LOCALE"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_LOCALE"]);
-		print(tbl.L["INFO_MSG_COMMAND_LOCALE"]);
-	elseif args == tbl.L["COMMAND_LOOT"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_LOOT"]);
-		print(tbl.L["INFO_MSG_COMMAND_LOOT"]);
-	elseif args == tbl.L["COMMAND_MAN"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_MAN"]);
-		print(tbl.L["INFO_MSG_COMMAND_MAN"]);
-	elseif args == tbl.L["COMMAND_REMOVE"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_REMOVE"]);
-		print(tbl.L["INFO_MSG_COMMAND_REMOVE"]);
-	elseif args == tbl.L["COMMAND_SEARCH"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_SEARCH"]);
-		print(tbl.L["INFO_MSG_COMMAND_SEARCH"]);
-	elseif args == tbl.L["COMMAND_VIEW"] then
-		print(tbl.L["ADDON_NAME"] .. tbl.L["COMMAND_VIEW"]);
-		print(tbl.L["INFO_MSG_COMMAND_VIEW"]);
-	end
-end
--- Synopsis: The player can toss commands to this function to learn more about them.