+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - "Scan on Loot" allows the player to get a more specific item link, instead of the generic item's link.
 - The addon now tracks the class and level of the character that looted the item. This information is available in the Search output.
+- A new command, progress, tells you how many items you've seen versus an approximation of what's in the game.
 
 ### Changed
 - The modes are now Debug, Normal, and Silent.
@@ -11,4 +12,6 @@
 ### Removed
 - No more import command.
 - No more discord command.
+- No more help command.
+- No more man command.
 - Items looted from containers are no longer considered by the addon.